Move head component into body
