+++ front/src/components/body/body.html
@@ -1,10 +1,29 @@
 <div class="container">
   <div class="columns">
-    <div class="column col-12">
-      <ge-head></ge-head>
-      <!-- ko template: { nodes: $componentTemplateNodes } -->
-      <!-- /ko -->
+    <div class="column col-1"></div>
+    <div class="column col-10">
+
+      <header class="navbar">
+        <section class="navbar-section">
+          <a href="#" class="navbar-brand">Geddit.LOL</a>
+
+          <!-- ko foreach: navItems -->
+          <a href="#" class="btn btn-link" data-bind="click: function() { $parent.currentItem($data) }, text: name"></a>
+          <!-- /ko -->
+        </section>
+
+        <section class="navbar-section">
+          <div class="input-group input-inline">
+            <input class="form-input" type="text" placeholder="search" />
+            <button class="btn btn-primary input-group-btn">Search</button>
+          </div>
+        </section>
+      </header>
+
+      <div data-bind="component: currentItem().component">
+      </div>
     </div>
+    <div class="column col-1"></div>
   </div>
 </div>
