--- front/src/components/head/head.html
@@ -1,19 +1,0 @@
-<header class="navbar">
-  <section class="navbar-section">
-    <a href="#" class="navbar-brand">Geddit.LOL</a>
-
-    <!-- ko foreach: navItems -->
-    <a href="#" class="btn btn-link" data-bind="click: $parent.currentItem($data), text: name"></a>
-    <!-- /ko -->
-  </section>
-
-  <section class="navbar-section">
-    <div class="input-group input-inline">
-      <input class="form-input" type="text" placeholder="search" />
-      <button class="btn btn-primary input-group-btn">Search</button>
-    </div>
-  </section>
-</header>
-<style>
-
-</style>
